+++ tests/test_rloo_trainer.py
@@ -25,7 +25,7 @@ from transformers import (
 )
 from transformers.utils import is_peft_available
 
-from trl import RLOOConfig, RLOOTrainer
+from trl.experimental.rloo import RLOOConfig, RLOOTrainer
 
 from .testing_utils import TrlTestCase, require_peft, require_vision, require_vllm
