+++ trl/__init__.py
@@ -88,8 +88,6 @@ _import_structure = {
         "PRMTrainer",
         "RewardConfig",
         "RewardTrainer",
-        "RLOOConfig",
-        "RLOOTrainer",
         "SFTConfig",
         "SFTTrainer",
         "WinRateCallback",
@@ -169,8 +167,6 @@ if TYPE_CHECKING:
         PRMTrainer,
         RewardConfig,
         RewardTrainer,
-        RLOOConfig,
-        RLOOTrainer,
         SFTConfig,
         SFTTrainer,
         WinRateCallback,
