+++ trl/trainer/__init__.py
@@ -62,8 +62,6 @@ _import_structure = {
     "prm_trainer": ["PRMTrainer"],
     "reward_config": ["RewardConfig"],
     "reward_trainer": ["RewardTrainer"],
-    "rloo_config": ["RLOOConfig"],
-    "rloo_trainer": ["RLOOTrainer"],
     "sft_config": ["SFTConfig"],
     "sft_trainer": ["SFTTrainer"],
     "utils": [
@@ -122,8 +120,6 @@ if TYPE_CHECKING:
     from .prm_trainer import PRMTrainer
     from .reward_config import RewardConfig
     from .reward_trainer import RewardTrainer
-    from .rloo_config import RLOOConfig
-    from .rloo_trainer import RLOOTrainer
     from .sft_config import SFTConfig
     from .sft_trainer import SFTTrainer
     from .utils import (
